Try fixing frame rate cap

+++ src/time.c
@@ -6,7 +6,7 @@
 /*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/31 14:34:58 by veilo             #+#    #+#             */
-/*   Updated: 2020/09/07 13:43:45 by ohakola          ###   ########.fr       */
+/*   Updated: 2020/09/07 14:55:50 by ohakola          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ void			cap_framerate(t_wolf3d *app)
 	float		time_per_frame;
 
 	time_per_frame = 1000.0 / (float)FPS;
-	if (app->delta_time < time_per_frame)
-		SDL_Delay((uint32_t)(time_per_frame - app->delta_time));
+	if ((uint32_t)app->delta_time < time_per_frame)
+		SDL_Delay((uint32_t)((uint32_t)time_per_frame - app->delta_time));
 }
 
 float			sin_time(t_wolf3d *app,
